JWT token added

+++ BankingApp/src/main/java/com/ak/BankingApp/service/impl/AccountServiceImpl.java
@@ -5,6 +5,7 @@ import com.ak.BankingApp.entity.Customer;
 import com.ak.BankingApp.repository.AccountRepository;
 import com.ak.BankingApp.repository.CustomerRepository;
 import com.ak.BankingApp.service.AccountService;
+import com.ak.BankingApp.service.JWTService;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
@@ -20,6 +21,9 @@ public class AccountServiceImpl implements AccountService {
     @Autowired
     private CustomerRepository customerRepository;
 
+    @Autowired
+    private JWTService jwtService;
+
     @Override
     public Account createAccount(Long customerId, Account account) {
         Optional<Customer> customer = customerRepository.findById(customerId);
@@ -32,19 +36,23 @@ public class AccountServiceImpl implements AccountService {
     }
 
     @Override
-    public List<Account> getAllAccounts() {
-        return accountRepository.findAll();
+    public List<Account> getAllAccounts(Long customerId) {
+        return accountRepository.findByCustomerId(customerId);
     }
 
     @Override
-    public Account getAccountById(Long id) {
-        return accountRepository.findById(id).orElse(null);
+    public Account getAccountById(Long id, Long customerId) {
+        Account account = accountRepository.findById(id).orElse(null);
+        if (account != null && account.getCustomer().getId().equals(customerId)){
+            return account;
+        }
+        throw new RuntimeException("Account not found or now owned by customer with Id : "+customerId);
     }
 
     @Override
-    public Account updateAccount(Long id, Account updatedAccount) {
+    public Account updateAccount(Long id, Account updatedAccount, Long customerId) {
         Optional<Account> account = accountRepository.findById(id);
-        if (account.isPresent()){
+        if (account.isPresent() && account.get().getCustomer().getId().equals(customerId)){
             Account existingAccount = account.get();
             existingAccount.setAccountNumber(updatedAccount.getAccountNumber());
             existingAccount.setAccountType(updatedAccount.getAccountType());
@@ -61,7 +69,12 @@ public class AccountServiceImpl implements AccountService {
     }
 
     @Override
-    public void deleteAccount(Long id) {
-        accountRepository.deleteById(id);
+    public void deleteAccount(Long id, Long customerId) {
+        Optional<Account> account = accountRepository.findById(id);
+        if (account.isPresent() && account.get().getCustomer().getId().equals(customerId)){
+            accountRepository.deleteById(id);
+        } else {
+            throw new RuntimeException("Account not found or now owned by customer with Id : "+customerId);
+        }
     }
 }
